Code clean-up before the final merge

+++ Tst/UnitTests/RegressionTests.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using Microsoft.Pc;
@@ -17,23 +16,14 @@ namespace UnitTests
         public void RemoveEmptyDirectories()
         {
             var baseDir = new DirectoryInfo(Constants.ScratchParentDirectory);
-            if (baseDir.Exists)
-            {
-                RecursiveRemoveEmptyDirectories(baseDir);
-            }
+            if (baseDir.Exists) RecursiveRemoveEmptyDirectories(baseDir);
         }
 
         private static void RecursiveRemoveEmptyDirectories(DirectoryInfo dir)
         {
-            foreach (DirectoryInfo subdir in dir.EnumerateDirectories())
-            {
-                RecursiveRemoveEmptyDirectories(subdir);
-            }
+            foreach (var subdir in dir.EnumerateDirectories()) RecursiveRemoveEmptyDirectories(subdir);
 
-            if (!dir.EnumerateFileSystemInfos().Any())
-            {
-                dir.Delete();
-            }
+            if (!dir.EnumerateFileSystemInfos().Any()) dir.Delete();
         }
     }
 
@@ -42,7 +32,7 @@ namespace UnitTests
     public class CompileOnlyRegressionTests
     {
         private static IEnumerable<TestCaseData> RegressionTestSuite =>
-            TestCaseLoader.FindTestCasesInDirectory(Constants.TestDirectory, new[] { "Pc" });
+            TestCaseLoader.FindTestCasesInDirectory(Constants.TestDirectory, new[] {"Pc"});
 
         [TestCaseSource(nameof(RegressionTestSuite))]
         public void TestRegressions(DirectoryInfo testDir)
@@ -59,31 +49,28 @@ namespace UnitTests
     public class PrtRegressionTests
     {
         private static IEnumerable<TestCaseData> RegressionTestSuite =>
-            TestCaseLoader.FindTestCasesInDirectory(Constants.TestDirectory, new[] { "Prt" });
-        
+            TestCaseLoader.FindTestCasesInDirectory(Constants.TestDirectory, new[] {"Prt"});
+
         [TestCaseSource(nameof(RegressionTestSuite))]
         public void TestRegressions(DirectoryInfo testDir)
         {
             var scratchDir = Directory.CreateDirectory(Constants.ScratchParentDirectory);
             var factory = new TestCaseFactory(scratchDir);
-            var testCaseC = factory.CreateTestCase(testDir,CompilerOutput.C);
+            var testCaseC = factory.CreateTestCase(testDir, CompilerOutput.C);
             TestAssertions.AssertTestCase(testCaseC);
         }
 
         [Test]
         public void TestTemp()
         {
-            DirectoryInfo tempDir = Directory.CreateDirectory(Path.Combine(Constants.ScratchParentDirectory, "TestTemp"));
+            var tempDir = Directory.CreateDirectory(Path.Combine(Constants.ScratchParentDirectory, "TestTemp"));
             var tempFilePath = new FileInfo(Path.Combine(Constants.SolutionDirectory, "tmp", "test.p"));
             var nativeFiles = tempFilePath.Directory.GetFiles("*.c");
 
-            if (!tempFilePath.Exists)
-            {
-                return;
-            }
+            if (!tempFilePath.Exists) return;
 
-            var testCase = new CompilerTestCase(tempDir, new PrtRunner(new[] { tempFilePath }, nativeFiles),
-                                                new ExecutionOutputValidator(0, null, null));
+            var testCase = new CompilerTestCase(tempDir, new PrtRunner(new[] {tempFilePath}, nativeFiles),
+                new ExecutionOutputValidator(0, null, null));
 
             TestAssertions.AssertTestCase(testCase);
         }
@@ -94,7 +81,7 @@ namespace UnitTests
     public class PrtSharpRegressionTests
     {
         private static IEnumerable<TestCaseData> RegressionTestSuite =>
-            TestCaseLoader.FindTestCasesInDirectory(Constants.TestDirectory, new[] { "Prt" });
+            TestCaseLoader.FindTestCasesInDirectory(Constants.TestDirectory, new[] {"Prt"});
 
         [TestCaseSource(nameof(RegressionTestSuite))]
         public void TestRegressions(DirectoryInfo testDir)
